Melhorando a estrutura da pagina home, e dando animação aos componentes
