+++ interface/src/pages/home/components/listInstances.tsx
@@ -1,0 +1,101 @@
+
+export const Instances = () => {
+    
+  
+    return (
+        <div className='info'>
+        <div className='header'>
+          <div className='header-col'>
+            <p className='header-col-title'>Instância</p>
+          </div>
+          <div className='header-col'>
+            <p className='header-col-title'>Interação</p>
+          </div>
+          <div className='header-col'>
+            <p className='header-col-title'>Operação</p>
+          </div>
+          <div className='header-col'>
+            <p className='header-col-title'>Data</p>
+          </div>
+          <div className='header-col'>
+            <p className='header-col-title'>Baixar</p>
+          </div>
+        </div>
+        <div className='container-table'>
+              <div className='container-data'>
+                <table>
+                  <tr>
+                    <td>IES</td>
+                    <td>0</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                  <tr>
+                    <td>IES</td>
+                    <td>1</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                  <tr>
+                    <td>IES</td>
+                    <td>0</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                  <tr>
+                    <td>IES</td>
+                    <td>1</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                  <tr>
+                    <td>IES</td>
+                    <td>0</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                  <tr>
+                    <td>IES</td>
+                    <td>1</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                  <tr>
+                    <td>IES</td>
+                    <td>0</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                  <tr>
+                    <td>IES</td>
+                    <td>1</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                  <tr>
+                    <td>IES</td>
+                    <td>0</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                  <tr>
+                    <td>IES</td>
+                    <td>1</td>
+                    <td>E</td>
+                    <td>10/12/23</td>
+                    <td>icone</td>
+                  </tr>
+                </table>
+              </div>
+         </div>
+      </div>
+    )}
